requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pillow==10.2.0
torch==2.2.2
torchvision==0.17.2
timm==0.9.16
pydantic==2.7.1
requests==2.31.0
python-multipart==0.0.9
